# api/swagger/fin-api.yaml
openapi: 3.0.3
info:
  title: FinTrack API Service
  version: 1.0.0
  description: REST API for managing user financial transactions.
servers:
  - url: http://localhost:8080
paths:
  /healthz:
    get:
      summary: Health check
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'
  /v1/users/{userID}/transactions:
    parameters:
      - $ref: '#/components/parameters/UserID'
    get:
      summary: List user transactions
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
    post:
      summary: Create transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionInput'
      responses:
        '201':
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/users/{userID}/transactions/{transactionID}:
    parameters:
      - $ref: '#/components/parameters/UserID'
      - $ref: '#/components/parameters/TransactionID'
    put:
      summary: Update transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionInput'
      responses:
        '200':
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Partially update transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionInput'
      responses:
        '200':
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete transaction
      responses:
        '204':
          description: Deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
components:
  parameters:
    UserID:
      name: userID
      in: path
      required: true
      schema:
        type: string
    TransactionID:
      name: transactionID
      in: path
      required: true
      schema:
        type: integer
        format: int64
  schemas:
    Health:
      type: object
      properties:
        status:
          type: string
          example: ok
    Transaction:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: string
        amount:
          type: number
          format: double
        category:
          type: string
        type:
          type: string
          enum: [income, expense]
        created_at:
          type: string
          format: date-time
    TransactionInput:
      type: object
      required: [amount, category, type]
      properties:
        amount:
          type: number
          format: double
        category:
          type: string
        type:
          type: string
          enum: [income, expense]
    Error:
      type: object
      properties:
        error:
          type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'